requirements.txt
Flask>=3.0,<4.0
pywebview>=4.4,<6.0
